packages/pdf/pyproject.toml
[tool.poetry]
name = "rpaframework-pdf"
version = "0.2.0"
description = "PDF utilities used by RPA Framework"
authors = [
	"RPA Framework <rpafw@robocorp.com>",
]
license = "Apache-2.0"
readme = "README.rst"

homepage = "https://rpaframework.org/"
documentation = "https://rpaframework.org/"
repository = "https://github.com/robocorp/rpaframework"

keywords = ["robotframework", "rpa", "automation"]
classifiers = [
	"License :: OSI Approved :: Apache Software License",
	"Development Status :: 3 - Alpha",
	"Operating System :: OS Independent",
	"Intended Audience :: Developers",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Software Development :: Libraries",
	"Framework :: Robot Framework :: Library",
	"Framework :: Robot Framework",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
]

packages = [
	{ include = "RPA", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.6"
rpaframework-core = "^5.2.0"
rpaframework-recognition = "^0.6.0"
robotframework = "^3.2.2"
robotframework-pythonlibcore = "^2.1.0"
"pdfminer.six" = "20201018"
pypdf2 = "^1.26.0"
fpdf = "^1.7.2"

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
flake8 = "^3.7.9"
pylint = "^2.4.4"
pytest = "^6.1.2"
mock = "^4.0.2"
pytest-cov = "^2.10.1"

[tool.black]
target-version = ["py36", "py37", "py38"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
addopts = "-v --cov=src --cov-report term-missing --cov-branch"
testpaths = ["tests"]
